+++ CHANGELOG.rst
@@ -43,13 +43,14 @@ Added
   `#112 <https://github.com/openego/eGon-data/issues/112>`_
 * Abstraction of hvmv and ehv substations
   `#9 <https://github.com/openego/eGon-data/issues/9>`_
+* Filter zensus being inside Germany and assign population to municipalities
+  `#7 <https://github.com/openego/eGon-data/issues/7>`_
 * RE potential areas data import
   `#124 <https://github.com/openego/eGon-data/issues/124>`_
 * Heat demand data import
   `#101 <https://github.com/openego/eGon-data/issues/101>`_
 * Demographic change integration
   `#47 <https://github.com/openego/eGon-data/issues/47>`_
-
 
 Changed
 -------